+++ src/headers/switcher-data-structs.hpp
@@ -7,7 +7,7 @@
 #include <QThread>
 #include <curl/curl.h>
 
-#include "scene-switch-target.hpp"
+#include "scene-group.hpp"
 #include "switch-audio.hpp"
 #include "switch-executable.hpp"
 #include "switch-file.hpp"
