+++ src/scene-group.cpp
